refactor(cli): extract shared helpers from service_order.go
Move parseEndDate, convertParamValue, and parseDurationString to cmd/helpers.go so they can be reused by upcoming workshop commands. No functional changes.
